change when music is played. Allow for specifying volume, pitch and name custom music.
Custom music is now played at the beginning of
the next tick instead of immediately. The Now Playing
action bar message is now displayed as well if
a custom display name of the sound is given.

It is also possible to specify the volume and pitch
of the music. These properties are optional and
default to 1.